+++ app/src/app/const/quiz.ts
@@ -1,13 +1,4 @@
-export interface Choice {
-  statement: string;
-  isAnswer: boolean;
-}
-
-export interface Quiz {
-  question: string;
-  choices: Choice[]
-  explanation: string;
-};
+import { Quiz } from "./quizInterface";
 
 export const QUIZ_DATA: Quiz[] = [
   {
