+++ app/src/app/const/quizInterface.ts
@@ -1,0 +1,10 @@
+export interface Choice {
+  statement: string;
+  isAnswer: boolean;
+}
+
+export interface Quiz {
+  question: string;
+  choices: Choice[]
+  explanation: string;
+};
