+++ app/src/app/pages/quiz/quiz.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { Quiz, Choice } from 'src/app/const/quiz';
+import { Quiz, Choice } from 'src/app/const/quizInterface';
 import { QuizService } from 'src/app/services/quiz.service';
 import { MatDialog } from '@angular/material/dialog';
 import { ConfirmDialogComponent } from '../../components/confirm-dialog/confirm-dialog.component';
